Force replace contrib package from pgxman
If contrib package is installed by a pg distribution and the user
installs the contrib package again, pgxman will override with the one
from itself.

One thing to note, if the pgxman overriding contrib package is removed,
the contrib extension will be removed, even though if it's first
avaiable from the pg distribution.